fix handling of CODE refs stored in stash
Perl has an optimization allowing code refs to be stored directly in the
stash, without an associated GLOB. For all stable perl releases, this
will only happen for the 'main' namespace. Some development releases
(and possibly future stable) releases will do this for all packages.

Account for code refs stored directly in the stash when trying to list
all of the methods for a package.